# api/swagger.yaml
swagger: "2.0"
info:
  title: Inventory Stock Management API
  description: API for checking product availability, reducing, and increasing product stock.
  version: 1.0.0
  contact:
    email: prajjawal.tiwari.0311@gmail.com

basePath: /inventory

paths:
  /availability/{productId}:
    get:
      summary: "Check product availability"
      description: "Get product details including stock availability."
      operationId: "getProductAvailability"
      parameters:
        - name: productId
          in: path
          description: ID of the product to check availability
          required: true
          type: string
      responses:
        200:
          description: Product availability details
          schema:
            $ref: '#/definitions/ProductAvailabilityResponse'
        404:
          description: Product not found
          schema:
            $ref: '#/definitions/ErrorModel'
        500:
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorModel'

  /stock/reduce/{productId}:
    put:
      summary: "Reduce product stock"
      description: "Reduce product stock after a purchase."
      operationId: "reduceProductQuantity"
      parameters:
        - name: productId
          in: path
          description: ID of the product to reduce stock
          required: true
          type: string
        - in: body
          name: body
          description: Quantity to reduce from product stock
          required: true
          schema:
            $ref: '#/definitions/ReduceStockRequest'
      responses:
        200:
          description: Stock reduced successfully
          schema:
            $ref: '#/definitions/StockChangeResponse'
        400:
          description: Insufficient stock or product not found
          schema:
            $ref: '#/definitions/ErrorModel'
        500:
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorModel'

  /stock/increase/{productId}:
    put:
      summary: "Increase product stock"
      description: "Increase product stock due to restock or return."
      operationId: "increaseProductQuantity"
      parameters:
        - name: productId
          in: path
          description: ID of the product to increase stock
          required: true
          type: string
        - in: body
          name: body
          description: Quantity to add to product stock
          required: true
          schema:
            $ref: '#/definitions/IncreaseStockRequest'
      responses:
        200:
          description: Stock increased successfully
          schema:
            $ref: '#/definitions/StockChangeResponse'
        404:
          description: Product not found
          schema:
            $ref: '#/definitions/ErrorModel'
        500:
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorModel'

definitions:
  ProductAvailabilityResponse:
    type: object
    properties:
      productId:
        type: string
        description: ID of the product
      name:
        type: string
        description: Name of the product
      description:
        type: string
        description: Detailed description of the product
      marketPrice:
        type: number
        description: Market price of the product
      ourPrice:
        type: number
        description: Price at which the product is sold
      quantity:
        type: integer
        description: Quantity available in stock
      imageUrl:
        type: string
        description: Image URL of the product

  ReduceStockRequest:
    type: object
    properties:
      quantityToReduce:
        type: integer
        description: Quantity to reduce from stock
    required:
      - quantityToReduce

  IncreaseStockRequest:
    type: object
    properties:
      quantityToAdd:
        type: integer
        description: Quantity to add to stock
    required:
      - quantityToAdd

  StockChangeResponse:
    type: object
    properties:
      message:
        type: string
        example: "Stock changed successfully"
      productId:
        type: string
        description: ID of the product
      remainingQuantity:
        type: integer
        description: Quantity remaining after reduction (if reducing stock)
      newQuantity:
        type: integer
        description: New quantity after increase (if increasing stock)

  ErrorModel:
    type: object
    properties:
      status:
        type: integer
        description: Error status code
      message:
        type: string
        description: Error message